requirements.txt
requests
pandas
beautifulsoup4
lxml
tqdm
python-dotenv
chromadb
openai